28/03/2021

+++ PFE-SERVER/src/main/java/com/pfe/demo/DAO/SuivisBullMedRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 import javax.transaction.Transactional;
 
-public interface SuivisBullMedRepository extends JpaRepository<SuivisBullMed,Long> {
+public interface SuivisBullMedRepository extends JpaRepository<SuivisBullMed, Integer> {
 
     SuivisBullMed findByNumBull(int numBull);
